backend/pyproject.toml
[tool.poetry]
name = "heartdiseasedetector"
version = "0.1.0"
description = "Web-Applications for Auscultation Data Analysis using Audio Spectrogram Transformer, fastapi, rabbitmq, postgresql and Streamlit"
authors = ["Vdmrl <vvluzin2004@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
transformers = "^4.49.0"
prometheus-fastapi-instrumentator = "^7.0.2"
redis = "^5.2.1"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.8.1"
fastapi-cache2 = "^0.2.2"
numpy = "<2.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
